+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Calendar Versioning](https://calver.org/).
 - Repo dependabot, code quality, and secret scanning in `openmod-tracker` source code repo.
 - Automated release process following monthly inventory auto-update.
 - Add the `get_scores.py` script to process the OpenSSF scores and a dedicated visualisation page in the dashboard (#217).
+- Return the ISO3 country code in `user_classifications.csv` instead of the country name (#96).
 
 ### Fixed